fix petii bug
